moved misc scripts
