+++ Assets/Prefabs/UI/SetupWizard/Kwaliteitsinstellingen.prefab
@@ -690,7 +690,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 1, y: 1}
-  m_AnchoredPosition: {x: 0, y: 0.000091360635}
+  m_AnchoredPosition: {x: 0, y: 0.000018024502}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &5158789987175011053
@@ -1656,7 +1656,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_SizeDelta.x
-      value: 810.75006
+      value: 3839.9998
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -1701,7 +1701,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 405.37503
+      value: 1919.9999
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -1873,7 +1873,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_SizeDelta.x
-      value: 810.75006
+      value: 3839.9998
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -1918,7 +1918,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 405.37503
+      value: 1919.9999
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2411,7 +2411,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_SizeDelta.x
-      value: 810.75006
+      value: 3839.9998
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2456,7 +2456,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 405.37503
+      value: 1919.9999
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2632,7 +2632,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_SizeDelta.x
-      value: 810.75006
+      value: 3839.9998
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2677,7 +2677,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 405.37503
+      value: 1919.9999
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2888,7 +2888,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_SizeDelta.x
-      value: 810.75006
+      value: 3839.9998
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
@@ -2933,7 +2933,7 @@ PrefabInstance:
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
       propertyPath: m_AnchoredPosition.x
-      value: 405.37503
+      value: 1919.9999
       objectReference: {fileID: 0}
     - target: {fileID: 2675933933807748605, guid: db07abc90ce404592bade1909e5c59e8,
         type: 3}
