--- Assets/Scripts/Interface.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9a7819de0b22e6b4aad49ea58e2ed13f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
