+++ Assets/Scripts/UI/InspectorPanels/DownloadInspector.cs
@@ -20,6 +20,7 @@ using System.Runtime.InteropServices;
 using GG.Extensions;
 using Netherlands3D.Coordinates;
 using Netherlands3D.Functionalities.AreaDownload;
+using Netherlands3D.Twin.Rendering;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
