+++ Assets/_Application/Rendering/RenderedThumbnail.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using UnityEngine.Rendering.Universal;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Rendering
 {
     public class RenderedThumbnail : MonoBehaviour
     {
