+++ Assets/_Application/Rendering/RenderedThumbnail.cs.meta
