+++ Assets/_Application/_Twin/SetBlockerLayer.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
 using UnityEngine;
 
 namespace Netherlands3D.Twin
